copilot review but wip
