+++ stellar-engine/pyproject.toml
@@ -25,7 +25,7 @@ authors = [
 keywords = ["stellar", "phlowers", "mechaphlowers"]
 
 dependencies = [
-    "mechaphlowers>=0.6.0rc2",
+    "mechaphlowers==0.6.0rc2",
 ]
 
 [project.urls]
